Split Nodes into Separate Files - Each of the nodes needs to go into a separate file now that they are ScriptableObjects otherwise Unity doesn't track assets correctly. - Now everything is in the BehaviorTree namespace. - Fixed a bug that was causing the tree to be corrupted when it is cloned.
