+++ AIController.cs
@@ -3,9 +3,11 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 
+using BehaviorTree;
+
 public class AIController : MonoBehaviour
 {
-	public BehaviorTree behavior;
+	public BehaviorTree.BehaviorTree behavior;
 
 	private Hashtable _data = new Hashtable();
 
@@ -28,6 +30,7 @@ public class AIController : MonoBehaviour
 			case NodeStatus.FAILURE:
 			case NodeStatus.SUCCESS:
 				Debug.Log( "Behavior tree completed with status " + status );
+				enabled = false; // stop trying to run the behavior tree.
 				break;
 		}
 	}
