+++ Nodes/Compositors.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 035a30d0370322a45a420605ea685d79
+folderAsset: yes
+timeCreated: 1426560017
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
