+++ Nodes/Compositors/Compositor.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: a60da3a19d5b11344b469c7157164a00
+timeCreated: 1426560017
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
