+++ Nodes/Compositors/Sequence.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: a26939a200edaad41abbcb9f56021558
+timeCreated: 1426560017
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
