+++ Nodes/Decorators/Fail.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: b3ec94de69c16394197dec3aead8a1ca
+timeCreated: 1426560017
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
