+++ Nodes/Decorators/RepeatUntilFail.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: dd9411b34900ece47b5fd6ef1023f2d5
+timeCreated: 1426558535
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
