+++ Nodes/Leaves/NullNode.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections;
+
+namespace BehaviorTree
+{
+	public class NullNode : LeafNode
+	{
+		public override void InitSelf( Hashtable data ) { }
+
+		public override NodeStatus TickSelf()
+		{
+			return NodeStatus.SUCCESS;
+		}
+	}
+}
