+++ Nodes/ParallelCompositors/SequenceParallel.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: b8dd5bdc02f772045ac2aa370c27a470
+timeCreated: 1426560017
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
