+++ Nodes/TreeNode.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 79cee2ef2ceb40f4483ae051e8c8cd97
+timeCreated: 1426558534
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
